% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zufast-package.R
\docType{package}
\name{zufast-package}
\alias{zufast}
\alias{zufast-package}
\title{zufast: Fast Portable 'C' Primitives for Parsing and Formatting Data}
\description{
A header-only library of small, fast, portable 'C99' primitives for converting external text and bytes into native values and back: correctly rounded number parsing, shortest round-trip number formatting, ISO 8601 dates and timestamps, booleans, universally unique identifiers (UUIDs), hexadecimal and Base64, 'XXH3' hashing (from 'xxHash'), half-precision floats, endian helpers and UTF-8 validation. Other packages use it through 'LinkingTo' alone; the 'R' functions exist to test, benchmark and demonstrate the 'C' layer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/zufast/}
  \item \url{https://github.com/pedrobtz/zufast}
  \item Report bugs at \url{https://github.com/pedrobtz/zufast/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Bjoern Hoehrmann (UTF-8 decoder state table in inst/include/zufast/utf8.h) [copyright holder]
  \item Daniel Lemire (fast_float, ported to C as ffc.h) [copyright holder]
  \item João Paulo Magalhaes (fast_float, ported to C as ffc.h) [copyright holder]
  \item The fast_float authors (fast_float, ported to C as ffc.h) [copyright holder]
  \item Koleman Nix (ffc.h, the C port of fast_float) [copyright holder]
  \item Ulf Adams (Ryu) [copyright holder]
  \item Yann Collet (xxHash) [copyright holder]
}

}
\keyword{internal}
