% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zufast-package.R
\docType{package}
\name{zufast-package}
\alias{zufast}
\alias{zufast-package}
\title{zufast: Fast Portable C Primitives for Parsing and Formatting Data}
\description{
A header-only library of small, fast, portable C99 primitives for converting external text and bytes into native values and back: correctly rounded number parsing, shortest round-trip number formatting, ISO 8601 dates and timestamps, booleans, UUIDs, hexadecimal and Base64, XXH3 hashing, half-precision floats, endian helpers and UTF-8 validation. Other packages use it through 'LinkingTo' alone; the R functions exist to test, benchmark and demonstrate the C layer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/zufast/}
  \item \url{https://github.com/pedrobtz/zufast}
  \item Report bugs at \url{https://github.com/pedrobtz/zufast/issues}
}

}
\author{
\strong{Maintainer}: Pedro Z \email{pedrobtz@gmail.com}

Authors:
\itemize{
  \item Pedro Z \email{pedrobtz@gmail.com}
}

Other contributors:
\itemize{
  \item Bjoern Hoehrmann (UTF-8 decoder state table in inst/include/zufast/utf8.h) [copyright holder]
}

}
\keyword{internal}
